# Escora-Admin-Portal-main/lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: "1.0.0"
servers:
  - url: /api
paths:
...
  /itineraries:
    get:
      operationId: listItineraries
      summary: List all itineraries
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Itinerary"
    post:
      operationId: createItinerary
      summary: Create an itinerary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItineraryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinerary"
  /itineraries/share/{token}:
    get:
      operationId: getItineraryByShareToken
      summary: Get itinerary by public share token
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinerary"
        "404":
          description: Not found
  /itineraries/{id}:
    get:
      operationId: getItinerary
      summary: Get itinerary by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinerary"
        "404":
          description: Not found
    patch:
      operationId: updateItinerary
      summary: Update an itinerary
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItineraryUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinerary"
    delete:
      operationId: deleteItinerary
      summary: Delete an itinerary
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Travel Categories ─────────────────────────────────────────
...
  /journal:
    get:
      operationId: listJournalPosts
      summary: List journal posts
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: published
          schema:
            type: boolean
        - in: query
          name: tag
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalPost"
    post:
      operationId: createJournalPost
      summary: Create a journal post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalPost"
  /journal/{id}:
    get:
      operationId: getJournalPost
      summary: Get journal post by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalPost"
        "404":
          description: Not found
    patch:
      operationId: updateJournalPost
      summary: Update a journal post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalPostUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalPost"
    delete:
      operationId: deleteJournalPost
      summary: Delete a journal post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Media ─────────────────────────────────────────────────────
...
  /media/{id}:
    delete:
      operationId: deleteMedia
      summary: Delete a media file
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Dashboard ─────────────────────────────────────────────────
...
components:
  schemas:
...
    Itinerary:
      type: object
      required: [id, shareToken, customerName, title, days, stays, pricing, currency, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        shareToken:
          type: string
        customerName:
          type: string
        customerEmail:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        title:
          type: string
        destination:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        days:
          type: string
        stays:
          type: string
        inclusions:
          type: ["string", "null"]
        exclusions:
          type: ["string", "null"]
        pricing:
          type: string
        currency:
          type: string
        status:
          type: string
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ItineraryInput:
      type: object
      required: [customerName, title]
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        title:
          type: string
        destination:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        coverImageUrl:
          type: string
        summary:
          type: string
        days:
          type: string
        stays:
          type: string
        inclusions:
          type: string
        exclusions:
          type: string
        pricing:
          type: string
        currency:
          type: string
        status:
          type: string
        createdBy:
          type: string
    ItineraryUpdate:
      type: object
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        title:
          type: string
        destination:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        coverImageUrl:
          type: string
        summary:
          type: string
        days:
          type: string
        stays:
          type: string
        inclusions:
          type: string
        exclusions:
          type: string
        pricing:
          type: string
        currency:
          type: string
        status:
          type: string
        createdBy:
          type: string

    # ── Order ────────────────────────────────────────────────────
...
    JournalPost:
      type: object
      required: [id, title, slug, category, published, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        category:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        authorName:
          type: ["string", "null"]
        authorRole:
          type: ["string", "null"]
        authorImageUrl:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        readTimeMinutes:
          type: ["integer", "null"]
        tags:
          type: ["string", "null"]
        published:
          type: boolean
        publishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    JournalPostInput:
      type: object
      required: [title, slug, category]
      properties:
        title:
          type: string
        slug:
          type: string
        category:
          type: string
        excerpt:
          type: string
        content:
          type: string
        authorName:
          type: string
        authorRole:
          type: string
        authorImageUrl:
          type: string
        imageUrl:
          type: string
        readTimeMinutes:
          type: integer
        published:
          type: boolean
        publishedAt:
          type: string
    JournalPostUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        category:
          type: string
        excerpt:
          type: string
        content:
          type: string
        authorName:
          type: string
        authorRole:
          type: string
        authorImageUrl:
          type: string
        imageUrl:
          type: string
        readTimeMinutes:
          type: integer
        published:
          type: boolean
        publishedAt:
          type: string

    # ── Media ────────────────────────────────────────────────────
...